Rename mpv_thumb module to video_thumb
Reflects support for multiple video thumbnail backends (ffmpeg and mpv)
