+++ src/pipeline/mod.rs
@@ -12,8 +12,8 @@ mod crop;
 mod finalize;
 mod rasterize;
 mod background;
+mod video_thumb;
 mod icc;
-mod mpv_thumb;
 
 pub type PipelineResult<T> = Result<T, PipelineError>;
