+++ src/pipeline/video_thumb.rs
